+++ byproxy/__init__.py
@@ -1,7 +1,7 @@
 """
-# byproxy
+# ByProxy
 
-byproxy is a simple package containing simple tools for proxy management and usage.
+ByProxy is a simple package contains simple tools for proxy management and usage. You can generate proxy dictionaries from a list of urls and gain information about the proxies.
 
 """
 from byproxy.proxy_maker import ProxyMaker